// src/lib/types.ts
export interface User {
    id: string
    name: string
    email?: string
    avatar: string
  }
  
  export interface Message {
    id: string
    content: string
    sender: {
      id: string
      name: string
      avatar: string
    }
    timestamp: Date
  }
  
  export interface Group {
    id: string
    name: string
    description: string
    createdAt: Date
    participants: User[]
    messages: Message[]
  }
  